+++ infra/variables.tf
@@ -59,3 +59,15 @@ variable "banshee_web_password" {
   type        = string
   sensitive   = true
 }
+
+variable "web_password" {
+  description = "Password for the Banshee web interface"
+  type        = string
+  sensitive   = true
+}
+
+variable "alert_recipients" {
+  description = "List of email addresses to receive alerts"
+  type        = list(string)
+  default     = ["gclark0812@gmail.com", "psmith1111@icloud.com"]
+}